Better error handling, html sanitizing, added reviewAdd

+++ controllers/Order.js
@@ -2,13 +2,22 @@
 
 var utils = require('../utils/writer.js');
 var Order = require('../service/OrderService');
+var session = require('../utils/SessionManager');
 
 
 //TODO IMPLEMENTARE TUTTO
 
 module.exports.orderAddressGET = function orderAddressGET (req, res, next) {
     var orderId = req.swagger.params['orderId'].value;
-    Order.orderAddressGET(orderId)
+
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+    Order.orderAddressGET(orderId, userId)
         .then(function (response) {
             utils.writeJson(res, response);
         })
@@ -18,13 +27,24 @@ module.exports.orderAddressGET = function orderAddressGET (req, res, next) {
 };
 
 module.exports.orderDetailsGET = function orderDetailsGET (req, res, next) {
+
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+
     var orderId = req.swagger.params['orderId'].value;
-    Order.orderDetailsGET(orderId)
+
+
+    Order.orderDetailsGET(orderId, userId)
         .then(function (response) {
             utils.writeJson(res, response);
         })
         .catch(function (response) {
-            utils.writeJson(res, response);
+            utils.writeJson(res, response, response.errorCode || 500);
         });
 };
 
@@ -37,7 +57,13 @@ module.exports.orderPlacePOST = function orderPlacePOST (req, res, next) {
     var addressStreetLine2 = req.swagger.params['addressStreetLine2'].value;
 
 
-    let userId = 19;//TODO MOCK VALUE DA CAMBIARE
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
     Order.orderPlacePOST(userId, addressStreetLine1,city,zip_code,province,country,addressStreetLine2)
         .then(function (response) {
             utils.writeJson(res, response);
